// src/types.d.ts
import type { GENRES } from './consts'

export type Genres = typeof GENRES[keyof typeof GENRES]

declare global {
  interface Array<T> {
    toSpliced: (start: number, deleteCount?:number, ...args?: any) => T[]
  }
}

export interface Books {
  library: Library[];
}

export interface Library {
  book: Book;
}

export interface Book {
  title:    string;
  pages:    number;
  genre:    string;
  cover:    string;
  synopsis: string;
  year:     number;
  ISBN:     string;
  author:   Author;
}

export interface Author {
  name:       string;
  otherBooks: string[];
}